New layout to demo app
